Using BigDecimal instead of Double for better precision

+++ src/main/java/io/elgayed/exchangerate/model/AmountConversion.java
@@ -1,5 +1,6 @@
 package io.elgayed.exchangerate.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -7,18 +8,18 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 @JsonPropertyOrder({"label", "publishedAt", "amount", "result"})
 public class AmountConversion extends ExchangeRate {
 	
-	private Double amount;
+	private BigDecimal amount;
 	
-	public AmountConversion(LocalDate publishedAt, String base, String currency, Double amount, Double result) {
+	public AmountConversion(LocalDate publishedAt, String base, String currency, BigDecimal amount, BigDecimal result) {
 		super(publishedAt, base, currency, result);
 		this.amount = amount;
 	}
 
-	public Double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Double amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
